Modify p13136.cpp
int to long long

+++ p13136.cpp
@@ -7,7 +7,7 @@ int		main(void)
 	ios::sync_with_stdio(false);
 	cin.tie(0);
 
-	int		r, c, n;
+	long long		r, c, n;
 
 	cin >> r; cin >> c; cin >> n;
 	if (r % n) r = r / n + 1;
